drivers: qcom: Fix OOB due to ts_whole_num type
The timer_value type is unsigned long long, the ts_whole_num type
is unsigned long that may cause OOB on 32bit.

+++ drivers/soc/qcom/boot_stats.c
@@ -319,7 +319,7 @@ static ssize_t bootkpi_reader(struct file *fp, struct kobject *obj,
 		size_t count)
 {
 	struct boot_marker *marker;
-	unsigned long ts_whole_num, ts_precision;
+	unsigned long long ts_whole_num, ts_precision;
 	static char *kpi_buf;
 	static int temp;
 	int ret = 0;
